+++ fpdf/fpdf.py
@@ -281,6 +281,9 @@ class FPDF:
         self.page_duration = 0  # optional pages display duration, cf. add_page()
         self.page_transition = None  # optional pages transition, cf. add_page()
         self.allow_images_transparency = True
+        # Do nothing by default. Allowed values: 'WARN', 'DOWNSCALE':
+        self.oversized_images = None
+        self.oversized_images_ratio = 2  # number of pixels per UserSpace point
         self._rotating = 0  # counting levels of nested rotation contexts
         self._markdown_leak_end_style = False
         # Only set if XMP metadata is added to the document:
@@ -2360,22 +2363,28 @@ class FPDF:
             name, img = hashlib.md5(name.getvalue()).hexdigest(), name
         else:
             name, img = str(name), name
-        if name not in self.images:
-            info = get_img_info(img or load_image(name), self.image_filter)
-            info["i"] = len(self.images) + 1
-            self.images[name] = info
+        info = self.images.get(name)
+        if info:
+            info["usages"] += 1
         else:
-            info = self.images[name]
+            if not img:
+                img = load_image(name)
+            info = get_img_info(img, self.image_filter)
+            info["i"] = len(self.images) + 1
+            info["usages"] = 1
+            self.images[name] = info
 
         # Automatic width and height calculation if needed
-        if w == 0 and h == 0:
-            # Put image at 72 dpi
+        if w == 0 and h == 0:  # Put image at 72 dpi
             w = info["w"] / self.k
             h = info["h"] / self.k
         elif w == 0:
             w = h * info["w"] / info["h"]
         elif h == 0:
             h = w * info["h"] / info["w"]
+
+        if self.oversized_images and info["usages"] == 1:
+            info = self._downscale_image(name, img, info, w, h)
 
         # Flowing mode
         if y is None:
@@ -2398,6 +2407,75 @@ class FPDF:
         if link:
             self.link(x, y, w, h, link)
 
+        return info
+
+    def _downscale_image(self, name, img, info, w, h):
+        width_in_pt, height_in_pt = w * self.k, h * self.k
+        lowres_name = f"lowres-{name}"
+        lowres_info = self.images.get(lowres_name)
+        if (
+            info["w"] > width_in_pt * self.oversized_images_ratio
+            and info["h"] > height_in_pt * self.oversized_images_ratio
+        ):
+            factor = (
+                min(info["w"] / width_in_pt, info["h"] / height_in_pt)
+                / self.oversized_images_ratio
+            )
+            if self.oversized_images.lower().startswith("warn"):
+                LOGGER.warning(
+                    "OVERSIZED: Image %s with size %.1fx%.1fpx is rendered at size %.1fx%.1fpt."
+                    " Set pdf.oversized_images = 'DOWNSCALE' to reduce embedded image size by a factor %.1f",
+                    name,
+                    info["w"],
+                    info["h"],
+                    width_in_pt,
+                    height_in_pt,
+                    factor,
+                )
+            elif self.oversized_images.lower() == "downscale":
+                dims = (
+                    round(width_in_pt * self.oversized_images_ratio),
+                    round(height_in_pt * self.oversized_images_ratio),
+                )
+                info["usages"] -= 1  # no need to embed the high-resolution image
+                if lowres_info:  # Great, we've already done the job!
+                    info = lowres_info
+                    if info["w"] * info["h"] < dims[0] * dims[1]:
+                        # The existing low-res image is too small, we need a bigger low-res image:
+                        info.update(
+                            get_img_info(
+                                img or load_image(name), self.image_filter, dims
+                            )
+                        )
+                        LOGGER.debug(
+                            "OVERSIZED: Updated low-res image with name=%s id=%d to dims=%s",
+                            lowres_name,
+                            info["i"],
+                            dims,
+                        )
+                    info["usages"] += 1
+                else:
+                    info = get_img_info(
+                        img or load_image(name), self.image_filter, dims
+                    )
+                    info["i"] = len(self.images) + 1
+                    info["usages"] = 1
+                    self.images[lowres_name] = info
+                    LOGGER.debug(
+                        "OVERSIZED: Generated new low-res image with name=%s dims=%s id=%d",
+                        lowres_name,
+                        dims,
+                        info["i"],
+                    )
+            else:
+                raise ValueError(
+                    f"Invalid value for attribute .oversized_images: {self.oversized_images}"
+                )
+        elif lowres_info:
+            # Embedding the same image in high-res after inserting it in low-res:
+            lowres_info.update(info)
+            del self.images[name]
+            info = lowres_info
         return info
 
     @contextmanager
@@ -2534,7 +2612,7 @@ class FPDF:
             # Page
             self._newobj()
             self._out("<</Type /Page")
-            self._out("/Parent 1 0 R")
+            self._out(f"/Parent {pdf_ref(1)}")
             page = self.pages[n]
             if page["duration"]:
                 self._out(f"/Dur {page['duration']}")
@@ -2543,7 +2621,7 @@ class FPDF:
             w_pt, h_pt = page["w_pt"], page["h_pt"]
             if w_pt != dw_pt or h_pt != dh_pt:
                 self._out(f"/MediaBox [0 0 {w_pt:.2f} {h_pt:.2f}]")
-            self._out("/Resources 2 0 R")
+            self._out(f"/Resources {pdf_ref(2)}")
 
             page_annots = self.annots[n]
             if page_annots:  # Annotations, e.g. links:
@@ -2599,7 +2677,7 @@ class FPDF:
             spid = self._struct_parents_id_per_page.get(self.n)
             if spid is not None:
                 self._out(f"/StructParents {spid}")
-            self._out(f"/Contents {self.n + 1} 0 R>>")
+            self._out(f"/Contents {pdf_ref(self.n + 1)}>>")
             self._out("endobj")
 
             # Page content
@@ -2613,7 +2691,7 @@ class FPDF:
         self.offsets[1] = len(self.buffer)
         self._out("1 0 obj")
         self._out("<</Type /Pages")
-        self._out("/Kids [" + " ".join(f"{3 + 2 * i} 0 R" for i in range(nb)) + "]")
+        self._out("/Kids [" + " ".join(pdf_ref(3 + 2 * i) for i in range(nb)) + "]")
         self._out(f"/Count {nb}")
         self._out(f"/MediaBox [0 0 {dw_pt:.2f} {dh_pt:.2f}]")
         self._out(">>")
@@ -2726,11 +2804,11 @@ class FPDF:
                 self._out(f"/BaseFont /{name}")
                 self._out(f"/Subtype /{my_type}")
                 self._out("/FirstChar 32 /LastChar 255")
-                self._out(f"/Widths {self.n + 1} 0 R")
-                self._out(f"/FontDescriptor {self.n + 2} 0 R")
+                self._out(f"/Widths {pdf_ref(self.n + 1)}")
+                self._out(f"/FontDescriptor {pdf_ref(self.n + 2)}")
                 if font["enc"]:
                     if "diff" in font:
-                        self._out(f"/Encoding {nf + font['diff']} 0 R")
+                        self._out(f"/Encoding {pdf_ref(nf + font['diff'])}")
                     else:
                         self._out("/Encoding /WinAnsiEncoding")
                 self._out(">>")
@@ -2765,7 +2843,7 @@ class FPDF:
                     s += " /FontFile"
                     if my_type != "Type1":
                         s += "2"
-                    s += f" {self.font_files[filename]['n']} 0 R"
+                    s += " " + pdf_ref(self.font_files[filename]["n"])
                 self._out(f"{s}>>")
                 self._out("endobj")
             elif my_type == "TTF":
@@ -2788,8 +2866,8 @@ class FPDF:
                 self._out("/Subtype /Type0")
                 self._out(f"/BaseFont /{fontname}")
                 self._out("/Encoding /Identity-H")
-                self._out(f"/DescendantFonts [{self.n + 1} 0 R]")
-                self._out(f"/ToUnicode {self.n + 2} 0 R")
+                self._out(f"/DescendantFonts [{pdf_ref(self.n + 1)}]")
+                self._out(f"/ToUnicode {pdf_ref(self.n + 2)}")
                 self._out(">>")
                 self._out("endobj")
 
@@ -2800,12 +2878,12 @@ class FPDF:
                 self._out("<</Type /Font")
                 self._out("/Subtype /CIDFontType2")
                 self._out(f"/BaseFont /{fontname}")
-                self._out(f"/CIDSystemInfo {self.n + 2} 0 R")
-                self._out(f"/FontDescriptor {self.n + 3} 0 R")
+                self._out(f"/CIDSystemInfo {pdf_ref(self.n + 2)}")
+                self._out(f"/FontDescriptor {pdf_ref(self.n + 3)}")
                 if font["desc"].get("MissingWidth"):
                     self._out(f"/DW {font['desc']['MissingWidth']}")
                 self._putTTfontwidths(font, ttf.maxUni)
-                self._out(f"/CIDToGIDMap {self.n + 4} 0 R")
+                self._out(f"/CIDToGIDMap {pdf_ref(self.n + 4)}")
                 self._out(">>")
                 self._out("endobj")
 
@@ -2882,7 +2960,7 @@ class FPDF:
                         v = v | 4
                         v = v & ~32  # SYMBOLIC font flag
                     self._out(f" /{kd} {v}")
-                self._out(f"/FontFile2 {self.n + 2} 0 R")
+                self._out(f"/FontFile2 {pdf_ref(self.n + 2)}")
                 self._out(">>")
                 self._out("endobj")
 
@@ -3015,11 +3093,15 @@ class FPDF:
         self._out(f"/W [{''.join(w)}]")
 
     def _putimages(self):
-        for info in sorted(self.images.values(), key=lambda info: info["i"]):
-            self._putimage(info)
-            del info["data"]
-            if "smask" in info:
-                del info["smask"]
+        for img_info in sorted(
+            self.images.values(), key=lambda img_info: img_info["i"]
+        ):
+            if img_info["usages"] == 0:
+                continue
+            self._putimage(img_info)
+            del img_info["data"]
+            if "smask" in img_info:
+                del img_info["smask"]
 
     def _putimage(self, info):
         if "data" not in info:
@@ -3034,7 +3116,7 @@ class FPDF:
         if info["cs"] == "Indexed":
             self._out(
                 f"/ColorSpace [/Indexed /DeviceRGB "
-                f"{len(info['pal']) // 3 - 1} {self.n + 1} 0 R]"
+                f"{len(info['pal']) // 3 - 1} {pdf_ref(self.n + 1)}]"
             )
         else:
             self._out(f"/ColorSpace /{info['cs']}")
@@ -3053,7 +3135,7 @@ class FPDF:
             self._out(f"/Mask [{trns}]")
 
         if self.allow_images_transparency and "smask" in info:
-            self._out(f"/SMask {self.n + 1} 0 R")
+            self._out(f"/SMask {pdf_ref(self.n + 1)}")
 
         self._out(f"/Length {len(info['data'])}>>")
         self._out(pdf_stream(info["data"]))
@@ -3086,10 +3168,14 @@ class FPDF:
             self._out("endobj")
 
     def _putxobjectdict(self):
-        i = [(x["i"], x["n"]) for x in self.images.values()]
-        i.sort()
-        for idx, n in i:
-            self._out(f"/I{idx} {n} 0 R")
+        img_ids = [
+            (img_info["i"], img_info["n"])
+            for img_info in self.images.values()
+            if img_info["usages"]
+        ]
+        img_ids.sort()
+        for idx, n in img_ids:
+            self._out(f"/I{idx} {pdf_ref(n)}")
 
     def _putresourcedict(self):
         # From section 10.1, "Procedure Sets", of PDF 1.7 spec:
@@ -3099,9 +3185,9 @@ class FPDF:
         # > (preferably, all of those listed in Table 10.1).
         self._out("/ProcSet [/PDF /Text /ImageB /ImageC /ImageI]")
         self._out("/Font <<")
-        f = [(x["i"], x["n"]) for x in self.fonts.values()]
-        f.sort()
-        for idx, n in f:
+        font_ids = [(x["i"], x["n"]) for x in self.fonts.values()]
+        font_ids.sort()
+        for idx, n in font_ids:
             self._out(f"/F{idx} {pdf_ref(n)}")
         self._out(">>")
         self._out("/XObject <<")